% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_maccs_pattern}
\alias{binary_maccs_pattern}
\title{MACCS SMARTS pattern that represent binary properties.}
\format{list of strings, SMARTS pattern}
\usage{
binary_maccs_pattern
}
\description{
List of SMARTS pattern extracted from the MACCS fingerprint definition that
are binary by nature. Those patterns can be calculated using 'substructure'
fingerprints with specified \code{substructure.pattern}. Please check
\code{\link[rcdk]{get.fingerprint}} for further details.
}
\keyword{datasets}
