% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor_tools.R
\name{calculate_descriptor_from_inchi}
\alias{calculate_descriptor_from_inchi}
\title{Calculate molecular descriptors from InChI}
\usage{
calculate_descriptor_from_inchi(smiles, desc_name, calc_3D_desc = FALSE,
  verbose = FALSE)
}
\arguments{
\item{desc_name}{string, name of the descriptor to calculate}

\item{calc_3D_desc}{boolean, indicating whether descriptors that require
3D coordinates should be calculated.}

\item{verbose}{boolean, should the function be verbose}

\item{inchi}{list of strings (1 x n_samples)}
}
\value{
named list
\itemize{
  \item \code{desc_vals}, named list with the descriptor vectors for each InChI
  \item \code{d_type}, string data type of the descriptor
  \item \code{desc_params}, named list of descriptor parameters used
}
}
\description{
This function takes a set of InChI and calculates for those the different
desired molecular descriptors provided by the CDK library through rcdk.
}
