% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor_tools.R
\name{calculate_descriptor_from_smiles}
\alias{calculate_descriptor_from_smiles}
\title{Calculate molecular descriptors from SMILES}
\usage{
calculate_descriptor_from_smiles(smiles, desc_name, skip_3D_desc = TRUE,
  verbose = FALSE)
}
\arguments{
\item{smiles}{list of strings (1 x n_samples)}

\item{desc_name}{string, name of the descriptor to calculate}

\item{skip_3D_desc}{boolean, indicating whether descriptors that require
3D coordinates should be skipped. Return value of the function will be
\code{NULL} in this case.}

\item{verbose}{boolean, should the function be verbose}
}
\value{
named list
\itemize{
  \item \code{desc_vals}, named list with the descriptor vectors for each SMILES
  \item \code{d_type}, string data type of the descriptor
  \item \code{desc_params}, named list of descriptor parameters used
}
}
\description{
This function takes a set of SMILES and calculates for those the different
desired molecular descriptors provided by the CDK library through rcdk.
}
