% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{calculate_fingerprints_from_smiles}
\alias{calculate_fingerprints_from_smiles}
\title{Calculate molecular fingerprints from smiles}
\usage{
calculate_fingerprints_from_smiles(smiles, fps_type = "maccs",
  fps_mode = "bit", verbose = FALSE, ...)
}
\arguments{
\item{smiles}{list of strings (1 x n_samples)}

\item{fps_type}{string, fingerprint definition}

\item{fps_mode}{string, which fingerprint type should be returned:
\itemize{
  \item "bit": binary fingerprint-vector
  \item "raw": (key, value)-pairs
  \itemize{
    \item key: the ftp definition, e.g. SMARTS string
    \item value: the number of occurances of fpt
  }
  \item "count": counting fingerprint-vector
}}

\item{verbose}{boolean, should the function be verbose}

\item{...}{parameters passed to the fingerprint calculation function}
}
\value{
list of \code{\link{fingerprint-class}} objects for the provided
    SMILES. \code{is.null(list[i])} is \code{true}, if the fingerprint for
    the corresponding SMILES could not be computed.
}
\description{
This function takes a set of smiles and calculates for those the different
sets of fingerprints. For a list of valid fingerprint definitions see
\code{\link{get.fingerpint}}.
}
