% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_maccs_pattern}
\alias{count_maccs_pattern}
\title{MACCS SMARTS pattern that can be counted}
\format{list of strings, SMARTS pattern}
\usage{
count_maccs_pattern
}
\description{
List of SMARTS pattern extracted from the MACCS fingerprint definition that
are countable. Those patterns are used of the "maccs" fingerprints with the
mode "count" should be calcualted, e.g. using
\code{\link{calculate_fingerprints_from_inchi}}.
}
\keyword{datasets}
