% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{fingerprints_to_matrix}
\alias{fingerprints_to_matrix}
\title{Construct a fingerprint matrix}
\usage{
fingerprints_to_matrix(fps, is_hashed = FALSE, sort_hash_keys = FALSE,
  add_colnames = FALSE)
}
\arguments{
\item{fps}{list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}

\item{is_hashed}{boolean, indicating whether the provided fingerprints are
hashed fingerprints, i.e. fps dimension might be different and direct
matching of the dimensions is not possible.}

\item{sort_hash_keys}{boolean, indicating whether the hash keys should be
sorted (using \code{\link[base]{sort}}), i.e. the columns of the
fingerprints matrix will be sorted.}

\item{add_colnames}{boolean, indicating whether the hash keys should be
added as column-names to the output matrix.}
}
\value{
fingerprint matrix (n_samples x n_fingerprints)
}
\description{
List of \code{\link[fingerprint]{fingerprint}} objects to matrix.
}
