% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{get_count_fingerprint_mask}
\alias{get_count_fingerprint_mask}
\title{Produce a binary mask to exclude molecular fingerprints}
\usage{
get_count_fingerprint_mask(count_fps)
}
\description{
Produce a binary mask to exclude molecular fingerprints
}
