% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor_tools.R
\name{get_descriptor_meta_data}
\alias{get_descriptor_meta_data}
\title{Metadata for CDK descriptors}
\usage{
get_descriptor_meta_data(desc_name)
}
\arguments{
\item{desc_name}{string, rcdk-like descriptor name}
}
\value{
named list containing the descriptor metadata:
\itemize{
  \item \code{is_3D}, boolean indicating whether the descriptor requires 3D coordinates
  \item \code{conf_func}, list of rcdk molecule-configuration functions to be
    applied before descriptor calculation
  \item \code{clone_atom_container}, boolean indicating, whether the atom
    container should be cloned before the descriptor is calculated.
  \item \code{desc_params}, vector of names lists that can be passed to the descriptor
  \item \code{d_type}, string either float or int(eger)
  \item \code{d_length}, length of the descriptor vector
}
}
\description{
This function returns the metadata, e.g. needs 3D, is sclar, etc., for the
specified descriptor. This information is mainly extracted from the CDK java
code.
}
