% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{get_fingerprint_dimensions}
\alias{get_fingerprint_dimensions}
\title{Determine the fingerprint dimension}
\usage{
get_fingerprint_dimensions(fps, check_all = FALSE, silent = FALSE)
}
\arguments{
\item{fps}{list of \code{\link[fingerprint]{fingerprint}} object}

\item{check_all}{logical, if \code{FALSE} only the first not null fingerprint
is used to determine the fingerprint dimension.}

\item{silent}{logical, if \code{TRUE} warnings are supressed.}
}
\value{
integer, dimension of the fingerprints or list of integers, if several
  different dimensions where found.
}
\description{
Given a list of fingerprints, e.g. the output of
\link{calculate_fingerprints_from_inchi}, the dimension of the fingerprints
is determined.
}
