% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{get_fingerprint_mask}
\alias{get_fingerprint_mask}
\title{Calculate mask to exclude molecular fingerpints given a set of criterias}
\usage{
get_fingerprint_mask(fps, remove_single_value = TRUE,
  remove_redundant = TRUE, remove_low_variance = TRUE,
  low_variance_tshd = 0.9)
}
\arguments{
\item{fps}{binary matrix, shape (n_samples x n_fingerprints)}

\item{remove_single_value}{binary, exclude fps with always the same value}

\item{remove_redundant}{binary, exclude redundant fps}

\item{remove_low_variance}{binary, exlcude fps with low variance}

\item{low_variance_tshd}{scalar, threshold for low variance (default = 0.9,
i.e. fps where >=90\% of the examples have the same value are removed)}
}
\value{
binary vector (1 x n_fingerpints):
\itemize{
  \item \code{TRUE}: keep fingerprint
  \item \code{FALSE}: exclude fingerprint
}
}
\description{
Given a binary fingerprint matrix a mask is calculated to exclude
fingerprints: \itemize{
  \item with the same value across the dataset
  \item with low variance, e.g. fp_i = 1 in 90\% of the data
  \item which are redunant, e.g. fp_i = 1 <=> fp_j = 1
}
where fp_i is the i'th fingerprint definition.
}
