% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{inchi2smiles}
\alias{inchi2smiles}
\title{Convert InChIs to SMILES}
\usage{
inchi2smiles(inchis)
}
\arguments{
\item{inchis}{strings, list of inchis}
}
\value{
list of smiles strings. The value is "NULL" if the corresponding
  InChI could not be converted or was an empty string.
}
\description{
Converts a list of InChIs to SMILES strings.
}
