% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logp_obabel_pattern}
\alias{logp_obabel_pattern}
\title{OpenBabel logp SMARTS pattern}
\format{list of strings, SMARTS pattern}
\usage{
logp_obabel_pattern
}
\description{
List of atom type SMARTS classification patterns extracted from [1].
}
\references{
\itemize{
  \item [1] \url{https://doi.org/10.1021/ci990307l}
}
}
\keyword{datasets}
