% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{setdiff_fingerprints}
\alias{setdiff_fingerprints}
\title{Calcualte the set-difference for two sets of fingerprints elementwise}
\usage{
setdiff_fingerprints(fps_A, fps_B)
}
\arguments{
\item{fps_A}{fps list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}

\item{fps_B}{fps list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}
}
\value{
list of \code{\link[fingerprint]{fingerprint-class}}.
}
\description{
Function to calcualte the set-difference for two sets of hashed fingerprints
for each example. This can be useful, if circular, i.e. ECFP or FCFP,
fingerprints are considered. For example if we want to know, which features
are unique to the ECFP6 fingerprints, we need to remove those ones of ECFP4.
}
\examples{
inchi <- "InChI=1S/C9H10O4/c10-7-3-1-6(2-4-7)5-8(11)9(12)13/h1-4,8,10-11H,5H2,(H,12,13)"
fps_ecfp4 <- calculate_fingerprints_from_inchi(
    inchi, fp_type = "circular", fp_mode = "count", circular.type="ECFP4")
fps_ecfp6 <- calculate_fingerprints_from_inchi(
    inchi, fp_type = "circular", fp_mode = "count", circular.type="ECFP6")

fps_diff <- setdiff_fingerprints(fps_ecfp6, fps_ecfp4)

}
