% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{write_fingerprint_mask_to_csv_file}
\alias{write_fingerprint_mask_to_csv_file}
\title{Store fingerprint mask into csv-file}
\usage{
write_fingerprint_mask_to_csv_file(mask, path)
}
\arguments{
\item{mask}{binary vector (1 x n_fingerpints), fingerprint mask calcualted
with \code{\link{get_fingerprint_mask}} or
\code{\link{get_count_fingerprint_mask}}.}

\item{path}{string, path to the output file}
}
\description{
Store fingerprint mask into csv-file
}
