% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{write_fingerprint_to_csv_file}
\alias{write_fingerprint_to_csv_file}
\title{Convert fingerprint list into a matrix and write it to a csv-file.}
\usage{
write_fingerprint_to_csv_file(fps, path)
}
\arguments{
\item{fps}{list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}

\item{path}{string, path to the output file}
}
\description{
List of \code{\link[fingerprint]{fingerprint}} objects is converted to a
matrix using \code{\link{fingerprint_to_matrix}} and subsequently
written into a csv-file.
}
