% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{write_fingerprint_to_json_file}
\alias{write_fingerprint_to_json_file}
\title{Convert fingerprint list into a JSON string and write it to a file.}
\usage{
write_fingerprint_to_json_file(fps, path, exclude_zero_fp = FALSE, ...)
}
\arguments{
\item{fps}{list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}

\item{path}{string, path to the output file}

\item{exclude_zero_fp}{logical, indicating whether fingerprints that have
value equal 0 should be excluded from the JSON string.}

\item{...}{parameters passed to down to \code{\link[jsonlite]{toJSON}}.}
}
\description{
List of \code{\link[fingerprint]{fingerprint}} objects is converted to a JSON
string. Subsequently this string is stored in a file using
\code{\link[jsonlite]{write_json}}.
}
