% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint_tools.R
\name{fingerprints_to_matrix}
\alias{fingerprints_to_matrix}
\title{Construct a fingerprint matrix}
\usage{
fingerprints_to_matrix(fps)
}
\arguments{
\item{fps}{list of \code{\link[fingerprint]{fingerprint}} objects
(1 x n_samples)}
}
\value{
fingerprint matrix (n_samples x n_fingerprints)
}
\description{
List of \code{\link[fingerprint]{fingerprint}} objects to matrix.
}
